Rename - club 디렉토리 내에 폴더 구조 변경(#19)
